requirements.txt
django-environ==0.11.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.32.0
google-cloud-discoveryengine==0.12.0
google-cloud-storage==2.18.0
langchain-google-community==1.0.7
langchain-google-genai==1.0.8
langchain-google-vertexai==1.0.8
gcsfs
google-cloud-pubsub==2.15.0